noir_ssa_fuzzer::builder
Struct FuzzerBuilder

pub struct FuzzerBuilder {
    pub(crate) builder: FunctionBuilder,
    pub(crate) runtime: RuntimeType,
}

Builder for generating fuzzed SSA functions Contains a FunctionBuilder and tracks the current numeric type being used

Fields
builder: FunctionBuilderruntime: RuntimeType
Implementations

impl FuzzerBuilder

pub fn new_acir(simplifying_enabled: bool) -> Self

Creates a new FuzzerBuilder in ACIR context

pub fn new_brillig(simplifying_enabled: bool) -> Self

Creates a new FuzzerBuilder in Brillig context

pub fn new_by_runtime(runtime: RuntimeType, simplifying_enabled: bool) -> Self

pub fn compile( self, compile_options: CompileOptions, ) -> Result<CompiledProgram, FuzzerBuilderError>

Compiles the built function into a CompiledProgram, to run it with nargo execute

pub fn insert_variable(&mut self, variable_type: SsaType) -> TypedValue

Inserts initial variables of the given type into the function

pub fn finalize_function(&mut self, return_value: &TypedValue)

Terminates main function block with the given value

pub fn insert_add_instruction_checked( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts an add instruction between two values

pub fn insert_sub_instruction_checked( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a subtract instruction between two values

pub fn insert_mul_instruction_checked( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a multiply instruction between two values

pub fn insert_div_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a divide instruction between two values

pub fn insert_mod_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a modulo instruction between two values

pub fn insert_not_instruction(&mut self, lhs: TypedValue) -> TypedValue

Inserts a not instruction for the given value

pub fn insert_cast(&mut self, value: TypedValue, cast_type: Type) -> TypedValue

Inserts a cast instruction

pub fn insert_eq_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts an equals comparison instruction between two values

pub fn insert_lt_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a less than comparison instruction between two values

pub fn insert_and_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a bitwise AND instruction between two values

pub fn insert_or_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a bitwise OR instruction between two values

pub fn insert_xor_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a bitwise XOR instruction between two values

pub fn insert_shl_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a left shift instruction between two values The right hand side is cast to 8 bits

pub fn insert_shr_instruction( &mut self, lhs: TypedValue, rhs: TypedValue, ) -> TypedValue

Inserts a right shift instruction between two values The right hand side is cast to 8 bits

pub fn insert_constant( &mut self, value: impl Into<FieldElement>, type_: NumericType, ) -> TypedValue

pub fn insert_constrain(&mut self, lhs: TypedValue, rhs: TypedValue)

pub fn get_current_block(&self) -> BasicBlockId

Gets the index of the entry block

pub fn switch_to_block(&mut self, block: BasicBlockId)

Switches the current block to the given block

pub fn insert_block(&mut self) -> BasicBlockId

Inserts a new basic block and returns its index

pub fn add_block_parameter( &mut self, block: BasicBlockId, typ: Type, ) -> TypedValue

Inserts a new parameter to the current SSA block and returns its value

b0() -> b0(new_parameter: parameter_type)

pub fn insert_return_instruction(&mut self, return_value: Id<Value>)

Inserts a return instruction with the given value

pub fn insert_jmp_instruction( &mut self, destination: BasicBlockId, arguments: Vec<TypedValue>, )

Inserts an unconditional jump to the given block with parameters

pub fn insert_jmpif_instruction( &mut self, condition: Id<Value>, then_destination: BasicBlockId, else_destination: BasicBlockId, )

Inserts a conditional jump based on the condition value

pub fn insert_add_to_memory(&mut self, lhs: TypedValue) -> TypedValue

pub fn insert_load_from_memory(&mut self, memory_addr: TypedValue) -> TypedValue

pub fn insert_set_to_memory( &mut self, memory_addr: TypedValue, value: TypedValue, )

pub fn new_function(&mut self, name: String, function_id: Id<Function>)

Creates a new function with the given name and id with inline type InlineType::Inline Sets the same runtime as the builder

pub fn insert_import(&mut self, function: Id<Function>) -> Id<Value>

Inserts an import function with the given function id

pub fn insert_call( &mut self, function: Id<Value>, arguments: &[TypedValue], result_type: Type, ) -> Id<Value>

Inserts a call to the given function with the given arguments and result type

pub fn insert_array(&mut self, elements: Vec<TypedValue>) -> TypedValue

Inserts a new array with the given elements and type

pub fn insert_vector(&mut self, elements: Vec<TypedValue>) -> TypedValue

pub fn insert_to_le_radix( &mut self, field_value: TypedValue, radix: u32, limb_count: u8, ) -> TypedValue

Inserts a to_le_radix intrinsic call that decomposes a field into little-endian radix representation

Arguments

field_value - The field value to decompose (must be a field type)

radix - The radix to use (must be a power of 2, between 2 and 256)

limb_count - Number of limbs in the result array

Returns

An array of u8 values representing the little-endian radix decomposition

pub fn insert_from_le_radix( &mut self, array: TypedValue, radix: u128, ) -> TypedValue

Inserts a from_le_radix hand-written function that composes a field from little-endian radix representation

Arguments

array - The array of u8 values to compose into a field

radix - The radix to use (must be a power of 2, between 2 and 256)

Returns

A field value representing the composed value

pub fn insert_blake2s_hash(&mut self, input: TypedValue) -> TypedValue

Inserts a blake2s hash intrinsic call

Arguments

input - The array of u8 values to hash

Returns

An array of u8 values representing the blake2s hash

pub fn insert_blake3_hash(&mut self, input: TypedValue) -> TypedValue

Inserts a blake3 hash intrinsic call

Arguments

input - The array of u8 values to hash

Returns

An array of u8 values representing the blake3 hash

pub fn insert_keccakf1600_permutation( &mut self, input: TypedValue, ) -> TypedValue

Inserts a keccakf1600 permutation intrinsic call

Arguments

input - The array of u64 values to permute

Returns

An array of u64 values representing the keccakf1600 permutation

pub fn insert_aes128_encrypt( &mut self, input: TypedValue, key: TypedValue, iv: TypedValue, ) -> TypedValue

Inserts a aes128_encrypt intrinsic call

Arguments

input - The array of u8 values to encrypt

key - The array of u8 values to use as key must be 16 bytes

iv - The array of u8 values to use as iv must be 16 bytes

Returns

An array of u8 values representing the encrypted input

pub fn insert_sha256_compression( &mut self, input: TypedValue, state: TypedValue, ) -> TypedValue

Inserts a sha256 compression intrinsic call

Arguments

input - The array of u32 values to compress must be 16 elements

state - The array of u32 values to use as state must be 8 elements

Returns

An array of u32 values representing the compressed input (8 elements)

fn insert_index_mod_array_length( &mut self, index: TypedValue, array: TypedValue, ) -> TypedValue

Inserts a modulo operation between the index and the array length Returns the id of the index modulo the array length

pub fn insert_array_get( &mut self, array: TypedValue, index: TypedValue, element_type: Type, safe_index: bool, ) -> TypedValue

Inserts an array get instruction

Index must be u32 If safe_index is true, index will be taken modulo the array length

pub fn insert_array_set( &mut self, array: TypedValue, index: TypedValue, value: TypedValue, safe_index: bool, ) -> TypedValue

Inserts an array set instruction

Index must be u32 If safe_index is true, index will be taken modulo the array length

pub fn base_scalar_mul( &mut self, scalar: Scalar, is_infinite: TypedValue, ) -> Point

Performs a curve point multiplication with a scalar

Arguments

scalar - The scalar value to multiply

is_infinite - The boolean value indicating if the resulting point is on the curve (setting it by ourself)

Returns

Point corresponding to value * G with overwritten on_curve value

pub fn create_point_from_scalar( &mut self, scalar: Scalar, is_infinite: TypedValue, ) -> Point

Creates a point from an affine x coordinate (scalar.lo, scalar.hi, is_infinite) Mostly invalid

Arguments

scalar - The scalar value to take coordinates from

is_infinite - The boolean value indicating if the resulting point is on the curve (setting it by ourself)

Returns

Point (scalar.lo, scalar.hi, is_infinite)

pub fn multi_scalar_mul( &mut self, points: Vec<Point>, scalars: Vec<Scalar>, predicate: bool, ) -> Point

pub fn point_add(&mut self, p1: Point, p2: Point, predicate: bool) -> Point

fn bytes_to_ssa_array(&mut self, vec: Vec<u8>) -> TypedValue

pub fn ecdsa_secp256r1( &mut self, pub_key_x: Vec<u8>, pub_key_y: Vec<u8>, hash: Vec<u8>, signature: Vec<u8>, predicate: bool, ) -> TypedValue

pub fn ecdsa_secp256k1( &mut self, pub_key_x: Vec<u8>, pub_key_y: Vec<u8>, hash: Vec<u8>, signature: Vec<u8>, predicate: bool, ) -> TypedValue